New functionality

+++ src/main/java/rocks/zipcode/atm/bank/Account.java
@@ -8,18 +8,21 @@ public abstract class Account {
     private AccountData accountData;
 
     public Account(AccountData accountData) {
+
         this.accountData = accountData;
     }
 
     public AccountData getAccountData() {
+
         return accountData;
     }
 
-    public void deposit(int amount) {
+    public void deposit(Double amount) {
+
         updateBalance(getBalance() + amount);
     }
 
-    public boolean withdraw(int amount) {
+    public boolean withdraw(Double amount) {
         if (canWithdraw(amount)) {
             updateBalance(getBalance() - amount);
             return true;
@@ -28,15 +31,17 @@ public abstract class Account {
         }
     }
 
-    protected boolean canWithdraw(int amount) {
+    protected boolean canWithdraw(Double amount) {
+
         return getBalance() >= amount;
     }
 
-    public int getBalance() {
+    public Double getBalance() {
+
         return accountData.getBalance();
     }
 
-    private void updateBalance(int newBalance) {
+    private void updateBalance(Double newBalance) {
         accountData = new AccountData(accountData.getId(), accountData.getName(), accountData.getEmail(),
                 newBalance);
     }
